GMAO_plot_docker/requirements.txt
cftime==1.3.0
cycler==0.10.0
eofs==1.4.0
kiwisolver==1.3.1
matplotlib==3.3.3
netCDF4==1.5.5.1
numpy==1.19.5
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.6.0
six==1.15.0
Cython
pyyaml
pandas
requests
schedule